Customizable mouse coordinate display location
Due to the placement of this year's elements, the mouse coordinate
display is not easily readable on top of the field image.

This contribution adds a function, `setMouseCoordinateDisplayPosition`,
which allows users to set a custom location on the field image to
display the mouse coordinates.

I've tested these changes with my code, and it works as expected, and I
didn't notice any issues.

To use the function, just call `setMouseCoordinateDisplayPosition` at
some point in your code before calling the `start()` function. Its x and
y parameters correspond to pixel locations relative to the top left
corner.